Bug#11752665: WINDOWS SERVER X64: SO MANY COMPILER WARNINGS
Patch #11:

This patch reduces the compiler warning count on Win64 by an additional
80, as reported by VS2013. The patch also removes a number of unnecessary casts.

+++ sql/rpl_rli.cc
@@ -193,7 +193,7 @@ Relay_log_info::~Relay_log_info()
 
     if(workers_copy_pfs.size())
     {
-      for (int i= workers_copy_pfs.size() - 1; i >= 0; i--)
+      for (int i= static_cast<int>(workers_copy_pfs.size()) - 1; i >= 0; i--)
         delete workers_copy_pfs[i];
       workers_copy_pfs.clear();
     }
@@ -2174,7 +2174,7 @@ bool Relay_log_info::read_info(Rpl_info_handler *from)
     overwritten by the second row later.
   */
   if (from->prepare_info_for_read() ||
-      from->get_info(group_relay_log_name, (size_t) sizeof(group_relay_log_name),
+      from->get_info(group_relay_log_name, sizeof(group_relay_log_name),
                      (char *) ""))
     DBUG_RETURN(TRUE);
 
@@ -2184,7 +2184,7 @@ bool Relay_log_info::read_info(Rpl_info_handler *from)
       *first_non_digit=='\0' && lines >= LINES_IN_RELAY_LOG_INFO_WITH_DELAY)
   {
     /* Seems to be new format => read group relay log name */
-    if (from->get_info(group_relay_log_name, (size_t) sizeof(group_relay_log_name),
+    if (from->get_info(group_relay_log_name, sizeof(group_relay_log_name),
                        (char *) ""))
       DBUG_RETURN(TRUE);
   }
@@ -2194,7 +2194,7 @@ bool Relay_log_info::read_info(Rpl_info_handler *from)
   if (from->get_info((ulong *) &temp_group_relay_log_pos,
                      (ulong) BIN_LOG_HEADER_SIZE) ||
       from->get_info(group_master_log_name,
-                     (size_t) sizeof(group_relay_log_name),
+                     sizeof(group_relay_log_name),
                      (char *) "") ||
       from->get_info((ulong *) &temp_group_master_log_pos,
                      (ulong) 0))
